agents/service-validation-agent/pyproject.toml
[project]
name = "service-validation-agent"
version = "0.1.0"
description = "Service validation agent for AWS SDP calibration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.3.1",
    "boto3>=1.42.51",
    "pydantic>=2.12.5",
    "strands-agents>=1.26.0",
]
